Module EndianBigstring

type bigstring = (char, Bigarray.int8_unsigned_elt, Bigarray.c_layout) Bigarray.Array1.t

module type EndianBigstringSig = sig ... end

Functions reading according to Big Endian byte order

module BigEndian : sig ... end

Functions reading according to Big Endian byte order without checking for overflow

module BigEndian_unsafe : sig ... end

Functions reading according to Big Endian byte order without checking for overflow

module LittleEndian : sig ... end

Functions reading according to Little Endian byte order

module LittleEndian_unsafe : sig ... end

Functions reading according to Big Endian byte order without checking for overflow

module NativeEndian : sig ... end

Functions reading according to machine endianness

module NativeEndian_unsafe : sig ... end

Functions reading according to machine endianness without checking for overflow